projects/01-flask-docker-ci/app/requirements.txt
flask
requests
python-dotenv
PyJWT==2.8.0
cryptography
    